add key to fragment children
