+++ packages/app/package.json
@@ -39,7 +39,7 @@
     "assert-ts": "^0.3.4",
     "effection": "^2.0.6",
     "history": "^5.0.0",
-    "platformscript": "^1.0.0-cl-exported-run.0",
+    "platformscript": "1.0.0-alpha.6",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router": "^6.3.0",
