+++ packages/app/src/components/PlatformScriptPage.tsx
@@ -1,120 +1,25 @@
-import React, { useCallback, useMemo, useRef, useState } from 'react';
-import type { ComponentType } from 'react';
+import React, { useCallback, useRef, useState } from 'react';
 import * as ps from 'platformscript';
-import { Button, Typography } from '@material-ui/core';
 import { MonacoEditor, YAMLEditor } from './yaml-editor/YamlEditor';
 import { useAsync } from 'react-use';
-import type { PSMap, PSValue } from 'platformscript';
-import { assert } from 'assert-ts';
-
-type TypographyProps = typeof Typography extends ComponentType<infer P>
-  ? P
-  : never;
-
-type Variant = TypographyProps['variant'];
+import type { PSValue } from 'platformscript';
+import { globals } from './globals';
 
 const DefultYaml = `$<>:
   - $Typography:
       variant: 'body1'
       children:
         "No links defined for this entity. You can add links to your entity YAML as shown in the highlighted example below:"
+  - $div:
+      className: 'whatever'
+      children:
+        "children"    
 `;
-
-export function lookup(key: string, map: PSMap): PSValue | void {
-  for (const entry of map.value.entries()) {
-    const [k, value] = entry;
-    if (k.value.toString() === key) {
-      return value;
-    }
-  }
-  return void 0;
-}
 
 export function PlatformScriptPage() {
   const editorRef = useRef<MonacoEditor>(null);
 
-  const globals = useMemo(() => ps.map({
-    alert: ps.fn(function* ({ arg, env }) {
-      const $arg = yield* env.eval(arg);
-
-      const message = String($arg.value);
-
-      // eslint-disable-next-line no-alert
-      window.alert(message);
-
-      return ps.boolean(true);
-    }),
-    '<>': ps.fn(function* ({ arg, env }) {
-      const $arg = yield* env.eval(arg);
-      
-      assert($arg.type === 'list', `a fragment must contain a list, found ${$arg.type}`);
-
-
-      const elements = $arg.value.map(value => value.value);
-
-      return ps.external(<>{elements}</>);
-    }),
-    Typography: ps.fn(function* ({ arg, env }) {
-      const $arg = yield* env.eval(arg);
-      let variant = '';
-      let children: any = '';
-
-      switch ($arg.type) {
-        case 'map':
-          children = lookup('children', $arg);
-
-          if (children) {
-            children = yield* env.eval(children);
-          }
-
-          variant = lookup('variant', $arg)?.value;
-
-          break;
-        default:
-          children = String($arg.type);
-      }
-
-      return ps.external(<Typography variant={variant as Variant}>{children.value}</Typography>);
-    },),
-    Button: ps.fn(function* ({ arg, env }) {
-      const $arg = yield* env.eval(arg);
-
-      let children = '';
-      let clickHandler = (): void => void 0;
-
-      switch ($arg.type) {
-        case 'string':
-          children = $arg.value;
-          break;
-        case 'map': {
-          children = lookup('text', $arg)?.value ?? '';
-
-          const psClickHandler = lookup('onClick', $arg);
-
-          if (psClickHandler) {
-            assert(psClickHandler.type === 'fn', `onClick must be a function but is ${psClickHandler.type}`);
-
-            clickHandler = () => {
-              ps.run(() => env.call(psClickHandler, {
-                arg: ps.boolean(true),
-                rest: ps.map({}),
-                env
-              }));
-            }
-          }
-
-          break;
-        }
-        default:
-          children = String($arg.type);
-      }
-
-
-      return ps.external(<Button onClick={clickHandler}>{children}</Button>);
-    },)
-  }), []);
-
-  const platformscript = useMemo(() => ps.createPlatformScript(globals), [globals]);
+  const platformscript = ps.createPlatformScript(globals);
 
   const [yaml, setYaml] = useState<string | undefined>(DefultYaml);
 
@@ -123,7 +28,7 @@ export function PlatformScriptPage() {
   }, []);
 
   const result = useAsync(async (): Promise<PSValue | undefined> => {
-    const program = platformscript.parse(yaml as string);
+    const program = ps.parse(yaml as string);
 
     const mod = await platformscript.eval(program);
 
@@ -131,7 +36,7 @@ export function PlatformScriptPage() {
   }, [yaml]);
 
   // eslint-disable-next-line @typescript-eslint/no-unused-expressions
-  result.error && console.log(result.error);
+  // result.error && console.log(result.error);
   
   return (
     <>
